requirements.txt
# LinkedIn AI Auto Job Applier - Requirements
# Python 3.10 or higher required

# Web Automation
selenium>=4.15.0
undetected-chromedriver>=3.5.4

# GUI Automation
PyAutoGUI>=0.9.54

# HTTP Requests
requests>=2.31.0

# YAML Configuration
PyYAML>=6.0.1

# Environment Variables
python-dotenv>=1.0.0

# AI/LLM Integration
openai>=1.3.0
google-generativeai>=0.3.0

# Web Framework (for Flask UI)
Flask>=3.0.0
flask-cors>=4.0.0

# Utilities
setuptools>=68.0.0
